delete directionModel
